feat(OH-123): 상수 수정
